Only add lobby distance filter if user does
Currently the distance filter defaults to close when no filter is
provided. This is unexpected. If the user did not give a filter no
filter should be applied. This change checks for a filter before the
dummy appid filter is created, and only adds the distance filter if a
filter was provided.