Messy proof of concept - introduced a new overload for SendMessage in Connection.cs that send data without creating a local copy.
Currently the user has to pin the data externally, and pass the GCHandle, next step is cleanup of message parameters and ensuring the method is fully sellf contained. The data handle is unpinned by SendMessage - this is bad custom as it did not pin it. planning to move the pin into the method.